requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
python-socketio==5.12.1
python-engineio==4.11.2
aiohttp>=3.8.5rc1
